manufacturing_optimization_pilot/requirements.txt
python-weka-wrapper3>=0.2.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0